Roles handling
To simplify things, I have added a sql query that is done when it's started the database that add the roles.
Than, if it's the first user added, automatically is the super admin.
If the mail end with @library365.it, the user is a librarian.
For the rest , it's a normal user